refactor: move textParsing logic into packages/api
