+++ packages/api/package.json
@@ -78,6 +78,7 @@
     "diff": "^7.0.0",
     "eventsource": "^3.0.2",
     "express": "^4.21.2",
+    "form-data": "^4.0.0",
     "js-yaml": "^4.1.0",
     "jsonwebtoken": "^9.0.0",
     "keyv": "^5.3.2",
