+++ packages/api/src/utils/index.ts
@@ -10,5 +10,6 @@ export * from './llm';
 export * from './math';
 export * from './openid';
 export * from './tempChatRetention';
+export * from './textParser';
 export { default as Tokenizer } from './tokenizer';
 export * from './yaml';
