+++ packages/api/src/utils/textParser.ts
@@ -1,0 +1,114 @@
+import fs from 'fs';
+import axios from 'axios';
+import FormData from 'form-data';
+import type { Request as ServerRequest } from 'express';
+import { generateShortLivedToken } from '../crypto/jwt';
+
+/**
+ * Attempts to parse text using RAG API, falls back to native text parsing
+ * @param {Object} params - The parameters object
+ * @param {Express.Request} params.req - The Express request object
+ * @param {Express.Multer.File} params.file - The uploaded file
+ * @param {string} params.file_id - The file ID
+ * @returns {Promise<{text: string, bytes: number, source: string}>}
+ */
+export async function parseText({
+  req,
+  file,
+  file_id,
+}: {
+  req: Pick<ServerRequest, 'user'> & {
+    user?: { id: string };
+  };
+  file: Express.Multer.File;
+  file_id: string;
+}): Promise<{ text: string; bytes: number; source: string }> {
+  if (!process.env.RAG_API_URL) {
+    console.debug('[parseText] RAG_API_URL not defined, falling back to native text parsing');
+    return parseTextNative(file);
+  }
+
+  try {
+    const healthResponse = await axios.get(`${process.env.RAG_API_URL}/health`, {
+      timeout: 5000,
+    });
+    if (healthResponse?.statusText !== 'OK' && healthResponse?.status !== 200) {
+      console.debug('[parseText] RAG API health check failed, falling back to native parsing');
+      return parseTextNative(file);
+    }
+  } catch (healthError) {
+    console.debug(
+      '[parseText] RAG API health check failed, falling back to native parsing',
+      healthError,
+    );
+    return parseTextNative(file);
+  }
+
+  try {
+    const jwtToken = generateShortLivedToken(req.user?.id || '');
+    const formData = new FormData();
+    formData.append('file_id', file_id);
+    formData.append('file', fs.createReadStream(file.path));
+
+    const formHeaders = formData.getHeaders();
+
+    // TODO: Actually implement referenced RAG API endpoint /parse-text
+    const response = await axios.post(`${process.env.RAG_API_URL}/parse-text`, formData, {
+      headers: {
+        Authorization: `Bearer ${jwtToken}`,
+        accept: 'application/json',
+        ...formHeaders,
+      },
+      timeout: 30000,
+    });
+
+    const responseData = response.data;
+    console.debug('[parseText] Response from RAG API', responseData);
+
+    if (!responseData.text) {
+      throw new Error('RAG API did not return parsed text');
+    }
+
+    return {
+      text: responseData.text,
+      bytes: Buffer.byteLength(responseData.text, 'utf8'),
+      source: 'rag_api',
+    };
+  } catch (error) {
+    console.warn('[parseText] RAG API text parsing failed, falling back to native parsing', error);
+    return parseTextNative(file);
+  }
+}
+
+/**
+ * Native JavaScript text parsing fallback
+ * Simple text file reading - complex formats handled by RAG API
+ * @param {Express.Multer.File} file - The uploaded file
+ * @returns {{text: string, bytes: number, source: string}}
+ */
+export function parseTextNative(file: Express.Multer.File): {
+  text: string;
+  bytes: number;
+  source: string;
+} {
+  try {
+    let text = '';
+
+    try {
+      text = fs.readFileSync(file.path, 'utf8');
+    } catch (readError) {
+      throw new Error(`Cannot read file as text: ${readError}`);
+    }
+
+    const bytes = Buffer.byteLength(text, 'utf8');
+
+    return {
+      text,
+      bytes,
+      source: 'native_js',
+    };
+  } catch (error) {
+    console.error('[parseTextNative] Error parsing file:', error);
+    throw new Error(`Failed to parse file: ${error}`);
+  }
+}
